django_backend/requirements.txt

Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-magic==0.4.27
Pillow==10.1.0
pdf2image==1.16.3
pytesseract==0.3.10
transformers==4.36.2
torch==2.1.1
numpy==1.26.2
pandas==2.1.3
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.0
